refactor smpp-repl
